Remove wrapped CWL doc example. (#4892)
* Remove wrapped CWL doc example.

* Patch missing links.

* Remove AWS dependant import/test from cwlTest.py.

* Missing @slow.

* Missing import.

* Make SimpleDB retry on EndpointConnectionError

* Linting.

* Satisfy MyPy more (new MyPy?)

---------
