+++ docs/cwl/running.rst
@@ -108,16 +108,6 @@ To run a CWL workflow in AWS with toil see :ref:`awscwl`.
 .. _secondaryFiles: http://www.commonwl.org/v1.0/CommandLineTool.html#CommandInputParameter
 .. _InitialWorkDirRequirement: http://www.commonwl.org/v1.0/CommandLineTool.html#InitialWorkDirRequirement
 
-Running CWL within Toil Scripts
-------------------------------------
-
-A CWL workflow can be run from a Toil Python workflow. However, this is not the :ref:`standard <cwl>` way to run
-CWL workflows with Toil and doing so comes at the cost of job efficiency. For some use cases, such as running one process on
-multiple files, it may be useful. For example, if you want to run a CWL workflow with 3 different input files specifying different
-samples inputs, it could look something like:
-
-.. literalinclude:: ../../src/toil/test/docs/scripts/tutorial_cwlexample.py
-
 Running CWL workflows with InplaceUpdateRequirement
 ---------------------------------------------------
