--- src/toil/test/docs/scripts/cwlExampleFiles/hello.cwl
@@ -1,14 +1,0 @@
-cwlVersion: v1.0
-class: CommandLineTool
-baseCommand: echo
-stdout: output.txt
-inputs:
-  - id: message
-    type: string
-    inputBinding:
-      position: 1
-outputs:
-  - id: output
-    type: File
-    outputBinding:
-      glob: output.txt
