--- src/toil/test/docs/scripts/cwlExampleFiles/hello1.yml
@@ -1,1 +1,0 @@
-message: Hello world!
