--- src/toil/test/docs/scripts/cwlExampleFiles/hello2.yml
@@ -1,1 +1,0 @@
-message: Hello world!
