--- src/toil/test/docs/scripts/cwlExampleFiles/hello3.yml
@@ -1,1 +1,0 @@
-message: Hello world!
